add scripts menu that list all supported scripts
