+++ index.html
@@ -2807,7 +2807,7 @@
                             >
                         </sp-radio-group>
                     </div>
-                    <div id="ultimateSDUpscalerContainer"></div>
+                    <div id="scriptsContainer"></div>
                 </div>
             </div>
         </div>
