+++ package.json
@@ -38,7 +38,12 @@
         "webpack-cli": "^5.1.1"
     },
     "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1"
+        "test": "echo \"Error: no test specified\" && exit 1",
+        "build": "cd ultimate_sd_upscaler && npx webpack --config webpack.config.js",
+
+  "watch": "cd ultimate_sd_upscaler && npx webpack --config webpack.config.js --watch"
+
+
     },
     "repository": {
         "type": "git",
