+++ ultimate_sd_upscaler/src/ultimate_sd_upscaler.tsx
@@ -9,140 +9,9 @@ import { SliderType, SpMenu, SpSliderWithLabel } from './elements'
 import * as sdapi from '../../sdapi_py_re'
 import { storeAnnotation } from 'mobx/dist/internal'
 
-let ui_config = {
-    tile_width: {
-        minimum: 0,
-        maximum: 2048,
-        step: 64,
-        label: 'Tile width',
-        value: 512,
-    },
-    tile_height: {
-        minimum: 0,
-        maximum: 2048,
-        step: 64,
-        label: 'Tile height',
-        value: 0,
-    },
-    mask_blur: {
-        minimum: 0,
-        maximum: 64,
-        step: 1,
-        label: 'Mask blur',
-        value: 8,
-    },
-    padding: { minimum: 0, maximum: 128, step: 1, label: 'Padding', value: 32 },
+import { ui_config } from './config'
 
-    seams_fix_denoise: {
-        label: 'Denoise',
-        minimum: 0,
-        maximum: 1,
-        step: 0.01,
-        value: 0.35,
-        visible: false,
-        interactive: true,
-    },
-
-    seams_fix_width: {
-        label: 'Width',
-        minimum: 0,
-        maximum: 128,
-        step: 1,
-        value: 64,
-        visible: false,
-        interactive: true,
-    },
-    seams_fix_mask_blur: {
-        label: 'Mask blur',
-        minimum: 0,
-        maximum: 64,
-        step: 1,
-        value: 4,
-        visible: false,
-        interactive: true,
-    },
-    seams_fix_padding: {
-        label: 'Padding',
-        minimum: 0,
-        maximum: 128,
-        step: 1,
-        value: 16,
-        visible: false,
-        interactive: true,
-    },
-    redraw_mode: {
-        label: 'Type',
-        choices: ['Linear', 'Chess', 'None'],
-        type: 'index',
-        value: 0,
-    },
-    save_upscaled_image: {
-        label: 'Upscaled',
-        value: true,
-    },
-    save_seams_fix_image: {
-        label: 'Seams fix',
-        value: false,
-    },
-
-    seams_fix_type: {
-        label: 'Type',
-        choices: [
-            'None',
-            'Band pass',
-            'Half tile offset pass',
-            'Half tile offset pass + intersections',
-        ],
-        type: 'index',
-        value: 0,
-    },
-
-    target_size_type: {
-        label: 'Target size type',
-        choices: [
-            'From img2img2 settings',
-            'Custom size',
-            'Scale from image size',
-        ],
-        type: 'index',
-        value: 2,
-    },
-
-    custom_width: {
-        label: 'Custom width',
-        minimum: 64,
-        maximum: 8192,
-        step: 64,
-        value: 2048,
-        visible: false,
-        interactive: true,
-    },
-    custom_height: {
-        label: 'Custom height',
-        minimum: 64,
-        maximum: 8192,
-        step: 64,
-        value: 2048,
-        visible: false,
-        interactive: true,
-    },
-    custom_scale: {
-        label: 'Scale',
-        minimum: 1,
-        maximum: 16,
-        step: 0.01,
-        value: 2,
-        visible: false,
-        interactive: true,
-    },
-
-    upscaler_index: {
-        label: 'Upscaler',
-        choices: [],
-        type: 'index',
-        value: 0,
-    },
-}
+export let script_name: string = 'Ultimate SD upscale'
 
 interface UltimateSDUpscalerData {
     _: string
@@ -224,14 +93,6 @@ class UltimateSDUpscalerStore {
     toJsFunc() {
         return toJS(this)
     }
-    orderedValues() {
-        const values = []
-        for (const key of script_args_ordered) {
-            // console.log(key, this.data[key])
-            values.push((this.data as any)[key])
-        }
-        return values
-    }
 }
 
 const configValues = Object.entries(ui_config).reduce(
@@ -246,13 +107,8 @@ export const ultimate_sd_upscaler_store = new UltimateSDUpscalerStore(
     default_values
 )
 
-// export let script_args: (string | number | boolean)[] = Object.values(
-//     ultimate_sd_upscaler_store.data
-// )
-export let script_name: string = 'Ultimate SD upscale'
-
 @observer
-class UltimateSDUpscalerForm extends React.Component<{
+export class UltimateSDUpscalerForm extends React.Component<{
     store: UltimateSDUpscalerStore
 }> {
     // slider1Ref = React.createRef<SpSliderWithLabel>()
@@ -293,16 +149,6 @@ class UltimateSDUpscalerForm extends React.Component<{
                     : new_index_value_pair['item']
             this.props.store.updateProperty(key, value)
         }
-    }
-    autoFillSliders = (newValue: any) => {
-        // scriptFormStore.setSlider1Value(newValue)
-        // if (this.slider1Ref.current) {
-        //     this.slider1Ref.current.setSliderValue(newValue)
-        // }
-        // if (this.slider2Ref.current) {
-        //     this.slider2Ref.current.setSliderValue(newValue)
-        // }
-        // scriptFormStore.setSlider2Value(newValue)
     }
 
     async getUpscalers() {
@@ -363,21 +209,6 @@ class UltimateSDUpscalerForm extends React.Component<{
         ))
         return (
             <div>
-                <sp-checkbox
-                    checked={this.props.store.is_active ? true : undefined}
-                    onClick={(event: React.ChangeEvent<HTMLInputElement>) => {
-                        this.props.store.setIsActive(event.target.checked)
-                    }}
-                >
-                    {'Activate'}
-                </sp-checkbox>
-                <button
-                    onClick={(event) =>
-                        this.handleSlider1ValueChange(
-                            this.props.store.test_value + 1
-                        )
-                    }
-                ></button>
                 <SpMenu
                     title="Stable Diffusion Upscalers"
                     items={this.state.sd_upscalers}
@@ -386,39 +217,8 @@ class UltimateSDUpscalerForm extends React.Component<{
                     id={'upscaler_index'}
                     onChange={this.handleMenuChange}
                 />
-                {/* <SpSliderWithLabel
-                    // ref={this.slider1Ref}
-                    // onSliderValueChangeCallback={this.handleSlider1ValueChange}
-                    id="tile_width"
-                    show-value={false}
-                    steps={64}
-                    out_min={ui_config.tile_width.minimum}
-                    out_max={ui_config.tile_width.maximum}
-                    // slider_value={this.props.store.test_value}
-                    output_value={parseInt(ui_config.tile_width.label)}
-                    title="this is a title"
-                    label={ui_config.tile_width.label}
-                    onSliderChange={this.handleSliderChange}
-                />
-
-                <p>test_value: {this.props.store.test_value}</p>
-                <SpSliderWithLabel
-                    // ref={this.slider2Ref}
-                    label={ui_config.tile_height.label}
-                    // onSliderValueChangeCallback={this.handleSlider2ValueChange}
-                />
-                 */}
                 {group_1_sliders}
                 {seamfix_sliders}
-                {/* <SpSliderWithLabel
-                    label={ui_config.mask_blur.label}
-                    // onSliderValueChangeCallback={this.handleSlider2ValueChange}
-                />
-                <SpSliderWithLabel
-                    label={ui_config.padding.label}
-                    // onSliderValueChangeCallback={this.handleSlider2ValueChange}
-                /> */}
-
                 <SpMenu
                     title=""
                     id={'target_size_type'}
@@ -437,7 +237,6 @@ class UltimateSDUpscalerForm extends React.Component<{
                     steps={0.01}
                     slider_type={SliderType.Float}
                 />
-
                 <SpMenu
                     title="Seams Fix Type"
                     id={'seams_fix_type'}
@@ -446,16 +245,6 @@ class UltimateSDUpscalerForm extends React.Component<{
                     onChange={this.handleMenuChange}
                     // style="width: 199px; margin-right: 5px"
                 />
-                {/* <SpSliderWithLabel label={ui_config.seams_fix_denoise.label} /> */}
-                {/* <SpSliderWithLabel label={ui_config.seams_fix_width.label} />
-                <SpSliderWithLabel
-                    label={ui_config.seams_fix_mask_blur.label}
-                    slider_value={this.props.store.data.seams_fix_mask_blur}
-                    id={'seams_fix_mask_blur'}
-                    onSliderChange={this.handleSliderChange}
-                />
-                <SpSliderWithLabel label={ui_config.seams_fix_padding.label} /> */}
-
                 <sp-checkbox
                     checked={
                         this.props.store.data.save_upscaled_image
@@ -486,43 +275,8 @@ class UltimateSDUpscalerForm extends React.Component<{
                 >
                     {ui_config.save_seams_fix_image.label}
                 </sp-checkbox>
-                <button onClick={() => this.autoFillSliders(50)}>
-                    Auto-fill sliders
-                </button>
-                <button onClick={this.handleUpdateItems}>
-                    Update Menu Items
-                </button>
+                \
             </div>
         )
     }
 }
-
-// @observer
-// class SliderValuesDisplay extends React.Component {
-//     render() {
-//         return (
-//             <div>
-//                 <p>
-//                     Slider test_value: {ultimate_sd_upscaler_store.test_value}
-//                 </p>
-//                 <p>
-//                     {/* Slider 1 value: {ultimate_sd_upscaler_store.data.tile_width} */}
-//                 </p>
-//                 <p>
-//                     Slider 2 value:{' '}
-//                     {ultimate_sd_upscaler_store.data.tile_height}
-//                 </p>
-//             </div>
-//         )
-//     }
-// }
-
-const domNode = document.getElementById('ultimateSDUpscalerContainer')!
-const root = ReactDOM.createRoot(domNode)
-root.render(
-    <React.StrictMode>
-        <UltimateSDUpscalerForm store={ultimate_sd_upscaler_store} />
-
-        {/* <SliderValuesDisplay /> */}
-    </React.StrictMode>
-)
